server/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.10.1
django-filter==2.4.0
coreapi==2.3.3
coreschema==0.0.4
psycopg2==2.9.1
requests==2.25.1
celery==4.4.7
redis==3.5.3
Crypto==1.4.1